GBS 4.1 update
GB studio updated to 4.1.x which separated the project file to several chunks. this also added the ability to modify the dialog boxes to a extent and add text noise to make a voice equivalent.
sound track was removed due to the file size being too big for github apparently.